difftest prompts you wether you want to do the xor diff

+++ gdsfactory/difftest.py
@@ -123,7 +123,6 @@ def run_xor(file1, file2, tolerance: int = 1, verbose: bool = False) -> None:
 def difftest(
     component: Component,
     test_name: Optional[str] = None,
-    xor: bool = True,
     dirpath: pathlib.Path = CONFIG["gdsdiff"],
 ) -> None:
     """Avoids GDS regressions tests on the GeometryDifference.
@@ -137,7 +136,6 @@ def difftest(
     Args:
         component: to test if it has changed.
         test_name: used to store the GDS file.
-        xor: runs xor if there is difference.
         dirpath: defaults to cwd refers to where the test is being invoked.
     """
     # containers function_name is different from component.name
@@ -168,7 +166,7 @@ def difftest(
         run_xor(str(ref_file), str(run_file), tolerance=1, verbose=False)
     except GeometryDifference as error:
         logger.error(error)
-        diff = gdsdiff(ref_file, run_file, name=test_name, xor=xor)
+        diff = gdsdiff(ref_file, run_file, name=test_name, xor=False)
         diff.write_gds(diff_file)
         diff.show(show_ports=False)
         print(
@@ -179,10 +177,20 @@ def difftest(
 
         try:
             val = input(
-                "Would you like to save current GDS as the new reference? [Y/n] "
+                "Save current GDS as new reference (Y) or show differences (d)? [Y/n/d]"
             )
             if val.upper().startswith("N"):
                 raise
+            xor = val.upper().startswith("D")
+            if xor:
+                diff = gdsdiff(ref_file, run_file, name=test_name, xor=xor)
+                diff.write_gds(diff_file)
+                diff.show(show_ports=False)
+
+                val = input("Save current GDS as the new reference (Y)? [Y/n]")
+                if val.upper().startswith("N"):
+                    raise
+
             logger.info(f"deleting file {str(ref_file)!r}")
             ref_file.unlink()
             shutil.copy(run_file, ref_file)
